Combine closely related integration test files
