+++ scripts/test/__init__.py
@@ -1,0 +1,4 @@
+import unittest
+
+# Don't show test docstrings when unittest is run with --verbose
+unittest.TestCase.shortDescription = lambda self: None
